feat(appliance): Add support for appliance type BASIC_FOR_DATABASES_VMS_*
When deploying a backup appliance, there should be 3 appliance types available.

Standard for Compute Engine VMs
Standard for VMware VMs and other databases or resources
Basic for VMware VMs and other databases or resources
3.a Minimal capacity persistent disk
3.b Standard Persistent Disk
3.c SSD Persistent Disk
The public documentation and Pantheon UI are inline with this.
Adding support in Terraform scripts as well.